hide post

+++ back/src/main/java/com/_blog/_blog/service/LikesService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import com._blog._blog.dto.IdDto;
 import com._blog._blog.dto.LikeDto;
+import com._blog._blog.exception.ForbiddenException;
 import com._blog._blog.exception.NotFoundException;
 import com._blog._blog.model.entity.AuthEntity;
 import com._blog._blog.model.entity.LikesEntity;
@@ -38,6 +39,9 @@ public class LikesService {
         PostsEntity post = postsRepo.findById(postIdDto.getId())
                 .orElseThrow(() -> new NotFoundException("post", "Post not found"));
 
+        if (post.getStatus().equals("Hide")) {
+            throw new ForbiddenException("authorization", "you are not authorization to like this post");
+        }
         Optional<LikesEntity> like = likesRepo.findByUserIdAndPostId(currentUser.getId(), post.getId());
         if (like.isPresent()) {
             likesRepo.delete(like.get());
